Refactor : 패키지 구조 수정
